Ensure that console.assert always calls logger with non-empty data
The specification says that  Logger should not perform anything
if data is empty. However, console.assert() should not be required
to provide an argument besides the condition to log something,
which matches what most implementations do.

This clarifies that section in the spec by providing an algorithm
as well as a recommendation of prefixing messages without destroying
formatting capabilities.