Handle input with no format specifiers in `Formatter`
This fixes #204.

As noted in that issue, several functions call `Formatter` without
checking that a format specifier is present, but the `Formatter`
algorithm assumed this was never the case. When called in this way,
`Formatter` would use the `result` variable without defining it.

This patch makes the following changes:

- Add a step 4 to the `Formatter` algorithm to explicitly handle a lack
of format specifier.

- Correct the indentation of what was step 3.8 of `Formatter`, so it is
now step 6.

- Remove what was step 4 from `Formatter`, since this is now redundant.

- Remove step 5 from the `Logger` algorithm, since this is now
redundant.

- Add a new step 1 to `Formatter`, which returns early if only one
argument was provided.